Update settings forms to use Type suffix
