+++ src/CSBill/SettingsBundle/Form/Type/SettingSectionType.php
@@ -1,0 +1,55 @@
+<?php
+
+/*
+ * This file is part of CSBill project.
+ *
+ * (c) 2013-2016 Pierre du Plessis <info@customscripts.co.za>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace CSBill\SettingsBundle\Form\Type;
+
+use CSBill\SettingsBundle\Manager\SettingsManager;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class SettingSectionType extends AbstractType
+{
+    /**
+     * @var SettingsManager
+     */
+    protected $manager;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        $this->manager = $options['manager'];
+
+        $settings = $this->manager->getSettings();
+
+        foreach ($settings as $section => $setting) {
+            $builder->add($section, SettingsType::class, ['section' => $this->manager->get($section)]);
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getBlockPrefix()
+    {
+        return 'settings';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver->setRequired(['manager']);
+    }
+}
